requirements.txt
PyMuPDF==1.26.3
numpy==2.3.2
google.generativeai==0.8.5
python-dotenv==1.1.1
faiss-cpu==1.12.0
SpeechRecognition==3.14.3
pyttsx3==2.99 
soundfile==0.13.1
streamlit==1.48.1
